samples/microsoft-agent-framework/multi-agent-eventgrid/pyproject.toml
[project]
name = "maf-multi-agent-eventgrid"
version = "0.1.0"
description = "Microsoft Agent Framework multi-agent pipeline with Azure Event Grid messaging"
requires-python = ">=3.11"

dependencies = [
    "azure-ai-agents>=1.0.0b4",
    "azure-ai-projects>=1.0.0b7",
    "azure-identity>=1.15.0",
    "azure-eventgrid>=4.20.0",
    "azure-storage-queue>=12.9.0",
    "structlog>=24.0.0",
    "python-dotenv>=1.0.0",
    "shared",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv.sources]
shared = { workspace = true }
